fix : countLike 타입 Long -> int

+++ NSU_FESTIVAL/src/main/java/com/example/nsu_festival/domain/booth/dto/BoothDetailDto.java
@@ -19,7 +19,7 @@ public class BoothDetailDto {
     private Long boothId;
     private String title;
     private String content;
-    private Long countLike;
+    private int countLike;
     private String area;
     private List<BoothCategory> boothCategories;
     private String boothImageUrl;
